wic: rawcopy: allow absolute path for source files, use basename when copying
When using an absolute path for source file in rawcopy `part` entry, we will hit
a non-obvious behavior of os.path.join() where it returns the last-occurring
absolute path as its result.

In such case, a temporary copy of the file made by rawcopy will be located
in the same directory as the source file instead of wic output directory.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>